models updated

+++ app/src/main/java/com/powerhouseweatherai/sumit/data/remote/dto/Wind.kt
@@ -1,7 +1,7 @@
 package com.powerhouseweatherai.sumit.data.remote.dto
 
 data class Wind(
-    val deg: Int?,
+    val deg: Long?,
     val gust: Double?,
     val speed: Double?
 )
