+++ app/src/main/java/com/powerhouseweatherai/sumit/domain/models/Main.kt
@@ -2,11 +2,11 @@ package com.powerhouseweatherai.sumit.domain.models
 
 data class Main(
     val feels_like: Double?,
-    val grnd_level: Int?,
-    val humidity: Int?,
-    val pressure: Int?,
-    val sea_level: Int?,
-    val temp: Int?,
+    val grnd_level: Long?,
+    val humidity: Long?,
+    val pressure: Long?,
+    val sea_level: Long?,
+    val temp: Double?,
     val temp_max: Double?,
     val temp_min: Double?
 )
