+++ app/src/main/java/com/powerhouseweatherai/sumit/domain/models/WeatherDetailResponse.kt
@@ -2,13 +2,13 @@ package com.powerhouseweatherai.sumit.domain.models
 
 data class WeatherDetailResponse(
     val base: String?,
-    val cod: Int?,
+    val cod: Long?,
     val coord: Coord?,
     val main: Main?,
     val name: String?,
     val sys: Sys?,
-    val timezone: Int?,
-    val visibility: Int?,
+    val timezone: Long?,
+    val visibility: Long?,
     val weather: List<Weather?>?,
     val wind: Wind?
 )
