+++ app/src/main/java/com/powerhouseweatherai/sumit/domain/models/Wind.kt
@@ -1,7 +1,7 @@
 package com.powerhouseweatherai.sumit.domain.models
 
 data class Wind(
-    val deg: Int?,
+    val deg: Long?,
     val gust: Double?,
     val speed: Double?
 )
